Time: 125 ms (77.54%) | Memory: 120.1 MB (50.78%) - LeetSync

+++ 3213-count-subarrays-where-max-element-appears-at-least-k-times/count-subarrays-where-max-element-appears-at-least-k-times.cpp
@@ -5,7 +5,7 @@ public:
         int count=0;
         int n=nums.size();
         long long int ans=0;
-        long long int maxi=*max_element(nums.begin(),nums.end());
+        int maxi=*max_element(nums.begin(),nums.end());
         while(j<n){
             if(nums[j]==maxi)count++;
